better build scripts
